New 'wasteline' django app
    remove wasteline information from the 'trak' app, including the data migrations
    that populates the database with initial DOT codes

    move test fixtures, and tests related to the waste line to our new 'wasteline' django app

    add urls to wasteline app

    add wasteline app models and views
