+++ server/apps/trak/models/__init__.py
@@ -6,4 +6,3 @@ from .contact_models import ManifestPhone
 from .handler_models import Handler, Transporter
 from .manifest_models import AdditionalInfo, Manifest
 from .signature_models import ESignature, PaperSignature, QuickerSign, Signer
-from .waste_models import DotLookup, WasteCode, WasteLine
