+++ server/apps/trak/services/__init__.py
@@ -5,4 +5,3 @@ from .manifest_services import (
     save_emanifest,
     update_manifest,
 )
-from .waste_services import get_dot_hazard_classes, get_dot_id_numbers, get_dot_shipping_names
