+++ server/apps/trak/views/__init__.py
@@ -5,10 +5,3 @@ from apps.trak.views.manifest_view import (
     SignManifestView,
     SyncSiteManifestView,
 )
-from apps.trak.views.waste_views import (
-    DotHazardClassView,
-    DotIdNumberView,
-    DotShippingNameView,
-    FederalWasteCodesView,
-    StateWasteCodesView,
-)
