+++ server/apps/wasteline/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class WastelineConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "apps.wasteline"
