+++ server/apps/wasteline/services.py
@@ -1,6 +1,6 @@
 from typing import Optional
 
-from apps.trak.models import DotLookup
+from apps.wasteline.models import DotLookup
 
 
 def get_dot_shipping_names(query: Optional[str]) -> list[str]:
@@ -21,4 +21,5 @@ def get_dot_hazard_classes(query: Optional[str]) -> list[str]:
 def get_dot_id_numbers(query: Optional[str]) -> list[str]:
     """Get a list of DOT ID Numbers (e.g., 'ID8000'), optionally filtered by a query parameter"""
     queryset = DotLookup.id_numbers.all().filter(value__icontains=query)[0:100]
+    print("queryset", queryset)
     return [option.value for option in queryset]
