+++ server/haztrak/urls.py
@@ -30,6 +30,7 @@ urlpatterns = [
                 path("", include("apps.core.urls")),
                 path("", include("apps.trak.urls")),
                 path("", include("apps.site.urls")),
+                path("", include("apps.wasteline.urls")),
                 path("schema/", SpectacularAPIView.as_view(), name="schema"),
                 path(
                     "schema/swagger-ui",
